Switching to mocha/chai/istanbul
