+++ package.json
@@ -16,7 +16,8 @@
     "url": "https://github.com/r3dm/thundercats"
   },
   "scripts": {
-    "test": "tape test/*.js"
+    "test": "mocha",
+    "cover": "istanbul cover _mocha"
   },
   "author": {
     "name": "Berkeley Martinez",
@@ -33,7 +34,8 @@
       "Name": "Berkeley Martinez",
       "email": "berkeley@r3dm.com",
       "url": "http://r3dm.com"
-    }, {
+    },
+    {
       "Name": "Nathan Leniz",
       "email": "terakilobyte@gmail.com",
       "url": "http://terakilobyte.com"
@@ -47,10 +49,15 @@
     "rx": "^2.3.14"
   },
   "devDependencies": {
-    "bluebird": "^2.5.3",
+    "q": "^1.2.0",
+    "chai": "^2.1.1",
+    "chai-as-promised": "^4.3.0",
+    "eslint-plugin-react": "^1.4.1",
+    "istanbul": "^0.3.7",
+    "mocha": "^2.2.1",
+    "mocha-istanbul": "^0.2.0",
     "rx": "^2.3.14",
     "sinon": "^1.11.1",
-    "tape": "^3.0.2",
-    "eslint-plugin-react": "^1.4.1"
+    "tape": "^3.0.2"
   }
 }
